+++ src/rendering/renderElement.js
@@ -6,7 +6,7 @@ export function renderElement(canvasElement) {
 
   const imageContext = getElementImageContext(canvasElement);
 
-  render(imageContext.data, imageContext.dimensions, imageContext.slice, imageContext.canvasImageData.data);
+  render(imageContext.imageData, imageContext.slice, imageContext.canvasImageData.data);
 
   imageContext.canvasContext.putImageData(imageContext.canvasImageData, 0, 0);